Adjacent locations are updated automatically on startup and when a new location is added
