+++ src/main/java/com/FlightPub/Controllers/HaversineCalculator.java
@@ -1,0 +1,25 @@
+package com.FlightPub.Controllers;
+
+public class HaversineCalculator {
+    // --- Haversine formula for finding distances between two coordinates ---
+    private static final int EARTH_RADIUS = 6371; // Approx Earth radius in KM
+
+    public static double distance(double startLat, double startLong,
+                                  double endLat, double endLong) {
+
+        double dLat  = Math.toRadians((endLat - startLat));
+        double dLong = Math.toRadians((endLong - startLong));
+
+        startLat = Math.toRadians(startLat);
+        endLat   = Math.toRadians(endLat);
+
+        double a = haversine(dLat) + Math.cos(startLat) * Math.cos(endLat) * haversine(dLong);
+        double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+
+        return EARTH_RADIUS * c; // <-- d
+    }
+
+    private static double haversine(double val) {
+        return Math.pow(Math.sin(val / 2), 2);
+    }
+}
